// Frontend/src/app/services/models/product-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProductDto {
  createdAt?: string;
  disableOutOfStock?: boolean;
  id?: string;
  note?: string;
  productDescription: string;
  productName: string;
  quantity: number;
  salePrice: number;
  shortDescription: string;
  sku?: string;
  slug: string;
}